requirements.txt
flask
flask-socketio
eventlet
python-dotenv
google-generativeai
pywebview
opencv-python 
mss 
numpy